Found bugs that needed to be fixed:
Prior to this update, cycling through tracks whilst playing would make the player would cycle two tracks as opposed to one. Now it cycles through one like it should.

Prior to this update, trying to pause the song whilst it was playing would instead just restart the song due to an 'if else' statement whithin the play_aud function supposed to automatically play the previous/next song when cycled through. There's now a completely seperate function for that so the two do no interfere